kimi_vl_a3b_thinking_2506/hf_tutorial/vllm_requirements_2.txt
aiohappyeyeballs==2.6.1+computecanada
aiohttp==3.12.13+computecanada
aiosignal==1.4.0+computecanada
airportsdata==20250706+computecanada
annotated_types==0.7.0+computecanada
anyio==3.7.1+computecanada
astor==0.8.1+computecanada
asttokens==3.0.0+computecanada
attrs==25.3.0+computecanada
blake3==1.0.4+computecanada
blobfile==3.0.0+computecanada
cachetools==6.1.0+computecanada
certifi==2025.8.3+computecanada
charset_normalizer==3.4.3+computecanada
click==8.2.1+computecanada
cloudpickle==3.1.1+computecanada
comm==0.2.2+computecanada
compressed_tensors==0.10.2+computecanada
contourpy==1.3.1+computecanada
cupy==13.3.0+computecanada
cycler==0.12.1+computecanada
debugpy==1.8.12+computecanada
decorator==5.1.1+computecanada
depyf==0.18.0+computecanada
dill==0.4.0+computecanada
diskcache==5.6.3+computecanada
distro==1.9.0+computecanada
dnspython==2.7.0+computecanada
einops==0.8.1+computecanada
email_validator==2.2.0+computecanada
executing==2.2.0+computecanada
fastapi==0.116.1+computecanada
fastapi_cli==0.0.8+computecanada
fastapi_cloud_cli==0.1.5+computecanada
fastrlock==0.8.3+computecanada
filelock==3.18.0+computecanada
flash_attn @ file:///lustre09/project/6062393/indrisch/vllm/kimi_vl_a3b_thinking_2506/hf_tutorial/flash_attn-2.8.1%2Bcu12torch2.7cxx11abiFALSE-cp312-cp312-linux_x86_64.whl#sha256=4bb004c17421b77b34b14f786710c6e4518e3fd3568c75407286517b4f50de34
fonttools==4.55.8+computecanada
frozenlist==1.7.0+computecanada
fsspec==2025.7.0+computecanada
gguf==0.17.1+computecanada
h11==0.16.0+computecanada
hf_xet==1.1.3+computecanada
httpcore==1.0.9+computecanada
httptools==0.6.4+computecanada
httpx==0.28.1+computecanada
huggingface_hub==0.34.4+computecanada
idna==3.10+computecanada
interegular==0.3.3+computecanada
ipykernel==6.29.5+computecanada
ipython==8.32.0+computecanada
ipywidgets==8.1.5+computecanada
jedi==0.19.2+computecanada
jinja2==3.1.6+computecanada
jiter==0.6.1+computecanada
jsonschema==4.25.0+computecanada
jsonschema_specifications==2025.4.1+computecanada
jupyter_client==8.6.3+computecanada
jupyter_core==5.7.2+computecanada
jupyterlab_widgets==3.0.13+computecanada
kiwisolver==1.4.8+computecanada
lark==1.2.2+computecanada
llguidance==0.7.14+computecanada
llvmlite==0.44.0+computecanada
lm_format_enforcer==0.10.12+computecanada
lxml==5.3.1+computecanada
markdown_it_py==3.0.0+computecanada
MarkupSafe==3.0.2+computecanada
matplotlib==3.10.0+computecanada
matplotlib_inline==0.1.7+computecanada
mdurl==0.1.2+computecanada
mistral_common==1.8.3+computecanada
mpmath==1.3.0+computecanada
msgpack==1.1.0+computecanada
msgspec==0.18.6+computecanada
multidict==6.6.3+computecanada
nest_asyncio==1.6.0+computecanada
networkx==3.5+computecanada
ninja==1.11.1+computecanada
numba==0.61.2+computecanada
numpy==2.2.2+computecanada
openai==1.90.0+computecanada
opencv_contrib_python==4.11.0
opencv_contrib_python_headless==4.11.0
opencv_python==4.11.0
opencv_python_headless==4.11.0
outlines==0.1.11+computecanada
outlines_core==0.1.26+computecanada
packaging==24.2+computecanada
pandas==2.2.3+computecanada
parso==0.8.4+computecanada
partial_json_parser==0.2.1.1.post6+computecanada
pexpect==4.9.0+computecanada
pillow==11.1.0+computecanada
Pillow_SIMD==9.5.0.post2+computecanada
platformdirs==4.3.6+computecanada
prometheus_client==0.22.1+computecanada
prometheus_fastapi_instrumentator==7.1.0+computecanada
prompt_toolkit==3.0.50+computecanada
propcache==0.3.2+computecanada
protobuf==6.31.1+computecanada
psutil==6.1.1+computecanada
ptyprocess==0.7.0+computecanada
pure_eval==0.2.3+computecanada
py_cpuinfo==9.0.0+computecanada
pybase64==1.4.1+computecanada
pycountry==24.6.1+computecanada
pycryptodomex==3.19.0+computecanada
pydantic==2.11.7+computecanada
pydantic_core==2.33.2+computecanada
pydantic_extra_types==2.10.5+computecanada
pygments==2.19.1+computecanada
pyparsing==3.2.1+computecanada
python_dateutil==2.9.0.post0+computecanada
python_dotenv==1.1.1+computecanada
python_json_logger==3.3.0+computecanada
python_multipart==0.0.20+computecanada
pytz==2025.1+computecanada
PyYAML==6.0.2+computecanada
pyzmq==26.2.1+computecanada
ray==2.43.0+computecanada
referencing==0.36.2+computecanada
regex==2024.11.6+computecanada
requests==2.32.4+computecanada
rich==14.1.0+computecanada
rich_toolkit==0.15.0+computecanada
rignore==0.6.2+computecanada
rpds_py==0.21.0+computecanada
safetensors==0.5.3+computecanada
scipy==1.15.1+computecanada
sentencepiece==0.2.0+computecanada
sentry_sdk==2.34.1+computecanada
setuptools==78.1.0+computecanada
shellingham==1.5.4+computecanada
six==1.17.0+computecanada
sniffio==1.3.1+computecanada
stack_data==0.6.3+computecanada
starlette==0.47.2+computecanada
sympy==1.13.3+computecanada
tiktoken==0.9.0+computecanada
tokenizers==0.21.1+computecanada
torch==2.7.0+computecanada
torchaudio==2.7.0+computecanada
torchvision==0.22.0+computecanada
tornado==6.4.2+computecanada
tqdm==4.67.1+computecanada
traitlets==5.14.3+computecanada
transformers==4.53.2+computecanada
triton==3.2.0+computecanada
typer==0.16.0+computecanada
typing_extensions==4.14.1+computecanada
typing_inspection==0.4.1+computecanada
tzdata==2025.1+computecanada
urllib3==2.5.0+computecanada
uvicorn==0.35.0+computecanada
uvloop==0.20.0+computecanada
vllm==0.9.2+computecanada
watchfiles==0.18.0+computecanada
wcwidth==0.2.13+computecanada
websockets==15.0.1+computecanada
widgetsnbextension==4.0.13+computecanada
xformers==0.0.30+computecanada
xgrammar==0.1.19+computecanada
yarl==1.20.1+computecanada
